backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
requests==2.32.3
pytest==8.3.5
pytest-django==4.12.0
pytest-cov==7.0.0
